+++ src/components/createProjectComponents/editor/MarkdownEditor.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as S from './MdEditorInput.styled';
+import * as S from './MarkdownEditor.styled';
 
 const MdEditorInput = ({ field }) => {
   return (
